# service/payment/api/payment/v1/payment_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.8.0
// - protoc             v7.34.1
// source: payment/v1/payment.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationPaymentCreateChannel = "/payment.v1.Payment/CreateChannel"
const OperationPaymentDeleteChannel = "/payment.v1.Payment/DeleteChannel"
const OperationPaymentGetChannel = "/payment.v1.Payment/GetChannel"
const OperationPaymentHandleCallback = "/payment.v1.Payment/HandleCallback"
const OperationPaymentListApiLogs = "/payment.v1.Payment/ListApiLogs"
const OperationPaymentListChannels = "/payment.v1.Payment/ListChannels"
const OperationPaymentProcessPayment = "/payment.v1.Payment/ProcessPayment"
const OperationPaymentQueryPayment = "/payment.v1.Payment/QueryPayment"
const OperationPaymentUpdateChannel = "/payment.v1.Payment/UpdateChannel"

type PaymentHTTPServer interface {
	CreateChannel(context.Context, *CreateChannelRequest) (*ChannelInfo, error)
	DeleteChannel(context.Context, *DeleteChannelRequest) (*DeleteChannelResponse, error)
	GetChannel(context.Context, *GetChannelRequest) (*ChannelInfo, error)
	HandleCallback(context.Context, *HandleCallbackRequest) (*HandleCallbackResponse, error)
	ListApiLogs(context.Context, *ListApiLogsRequest) (*ListApiLogsResponse, error)
	ListChannels(context.Context, *ListChannelsRequest) (*ListChannelsResponse, error)
	ProcessPayment(context.Context, *ProcessPaymentRequest) (*PaymentResponse, error)
	QueryPayment(context.Context, *QueryPaymentRequest) (*PaymentResponse, error)
	UpdateChannel(context.Context, *UpdateChannelRequest) (*ChannelInfo, error)
}

func RegisterPaymentHTTPServer(s *http.Server, srv PaymentHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/payment/channels", _Payment_ListChannels0_HTTP_Handler(srv))
	r.GET("/api/v1/payment/channels/{id}", _Payment_GetChannel0_HTTP_Handler(srv))
	r.POST("/api/v1/payment/channels", _Payment_CreateChannel0_HTTP_Handler(srv))
	r.PUT("/api/v1/payment/channels/{id}", _Payment_UpdateChannel0_HTTP_Handler(srv))
	r.DELETE("/api/v1/payment/channels/{id}", _Payment_DeleteChannel0_HTTP_Handler(srv))
	r.POST("/api/v1/payment/process", _Payment_ProcessPayment0_HTTP_Handler(srv))
	r.GET("/api/v1/payment/query/{order_no}", _Payment_QueryPayment0_HTTP_Handler(srv))
	r.GET("/api/v1/payment/logs", _Payment_ListApiLogs0_HTTP_Handler(srv))
	r.POST("/api/v1/payment/callback", _Payment_HandleCallback0_HTTP_Handler(srv))
}

func _Payment_ListChannels0_HTTP_Handler(srv PaymentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListChannelsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentListChannels)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListChannels(ctx, req.(*ListChannelsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListChannelsResponse)
		return ctx.Result(200, reply)
	}
}

func _Payment_GetChannel0_HTTP_Handler(srv PaymentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetChannelRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentGetChannel)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetChannel(ctx, req.(*GetChannelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ChannelInfo)
		return ctx.Result(200, reply)
	}
}

func _Payment_CreateChannel0_HTTP_Handler(srv PaymentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateChannelRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentCreateChannel)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateChannel(ctx, req.(*CreateChannelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ChannelInfo)
		return ctx.Result(200, reply)
	}
}

func _Payment_UpdateChannel0_HTTP_Handler(srv PaymentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateChannelRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentUpdateChannel)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateChannel(ctx, req.(*UpdateChannelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ChannelInfo)
		return ctx.Result(200, reply)
	}
}

func _Payment_DeleteChannel0_HTTP_Handler(srv PaymentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteChannelRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentDeleteChannel)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteChannel(ctx, req.(*DeleteChannelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteChannelResponse)
		return ctx.Result(200, reply)
	}
}

func _Payment_ProcessPayment0_HTTP_Handler(srv PaymentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ProcessPaymentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentProcessPayment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ProcessPayment(ctx, req.(*ProcessPaymentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PaymentResponse)
		return ctx.Result(200, reply)
	}
}

func _Payment_QueryPayment0_HTTP_Handler(srv PaymentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in QueryPaymentRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentQueryPayment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.QueryPayment(ctx, req.(*QueryPaymentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PaymentResponse)
		return ctx.Result(200, reply)
	}
}

func _Payment_ListApiLogs0_HTTP_Handler(srv PaymentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListApiLogsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentListApiLogs)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListApiLogs(ctx, req.(*ListApiLogsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListApiLogsResponse)
		return ctx.Result(200, reply)
	}
}

func _Payment_HandleCallback0_HTTP_Handler(srv PaymentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in HandleCallbackRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentHandleCallback)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.HandleCallback(ctx, req.(*HandleCallbackRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*HandleCallbackResponse)
		return ctx.Result(200, reply)
	}
}

type PaymentHTTPClient interface {
	CreateChannel(ctx context.Context, req *CreateChannelRequest, opts ...http.CallOption) (rsp *ChannelInfo, err error)
	DeleteChannel(ctx context.Context, req *DeleteChannelRequest, opts ...http.CallOption) (rsp *DeleteChannelResponse, err error)
	GetChannel(ctx context.Context, req *GetChannelRequest, opts ...http.CallOption) (rsp *ChannelInfo, err error)
	HandleCallback(ctx context.Context, req *HandleCallbackRequest, opts ...http.CallOption) (rsp *HandleCallbackResponse, err error)
	ListApiLogs(ctx context.Context, req *ListApiLogsRequest, opts ...http.CallOption) (rsp *ListApiLogsResponse, err error)
	ListChannels(ctx context.Context, req *ListChannelsRequest, opts ...http.CallOption) (rsp *ListChannelsResponse, err error)
	ProcessPayment(ctx context.Context, req *ProcessPaymentRequest, opts ...http.CallOption) (rsp *PaymentResponse, err error)
	QueryPayment(ctx context.Context, req *QueryPaymentRequest, opts ...http.CallOption) (rsp *PaymentResponse, err error)
	UpdateChannel(ctx context.Context, req *UpdateChannelRequest, opts ...http.CallOption) (rsp *ChannelInfo, err error)
}

type PaymentHTTPClientImpl struct {
	cc *http.Client
}

func NewPaymentHTTPClient(client *http.Client) PaymentHTTPClient {
	return &PaymentHTTPClientImpl{client}
}

func (c *PaymentHTTPClientImpl) CreateChannel(ctx context.Context, in *CreateChannelRequest, opts ...http.CallOption) (*ChannelInfo, error) {
	var out ChannelInfo
	pattern := "/api/v1/payment/channels"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentCreateChannel))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PaymentHTTPClientImpl) DeleteChannel(ctx context.Context, in *DeleteChannelRequest, opts ...http.CallOption) (*DeleteChannelResponse, error) {
	var out DeleteChannelResponse
	pattern := "/api/v1/payment/channels/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPaymentDeleteChannel))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PaymentHTTPClientImpl) GetChannel(ctx context.Context, in *GetChannelRequest, opts ...http.CallOption) (*ChannelInfo, error) {
	var out ChannelInfo
	pattern := "/api/v1/payment/channels/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPaymentGetChannel))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PaymentHTTPClientImpl) HandleCallback(ctx context.Context, in *HandleCallbackRequest, opts ...http.CallOption) (*HandleCallbackResponse, error) {
	var out HandleCallbackResponse
	pattern := "/api/v1/payment/callback"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentHandleCallback))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PaymentHTTPClientImpl) ListApiLogs(ctx context.Context, in *ListApiLogsRequest, opts ...http.CallOption) (*ListApiLogsResponse, error) {
	var out ListApiLogsResponse
	pattern := "/api/v1/payment/logs"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPaymentListApiLogs))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PaymentHTTPClientImpl) ListChannels(ctx context.Context, in *ListChannelsRequest, opts ...http.CallOption) (*ListChannelsResponse, error) {
	var out ListChannelsResponse
	pattern := "/api/v1/payment/channels"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPaymentListChannels))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PaymentHTTPClientImpl) ProcessPayment(ctx context.Context, in *ProcessPaymentRequest, opts ...http.CallOption) (*PaymentResponse, error) {
	var out PaymentResponse
	pattern := "/api/v1/payment/process"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentProcessPayment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PaymentHTTPClientImpl) QueryPayment(ctx context.Context, in *QueryPaymentRequest, opts ...http.CallOption) (*PaymentResponse, error) {
	var out PaymentResponse
	pattern := "/api/v1/payment/query/{order_no}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPaymentQueryPayment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *PaymentHTTPClientImpl) UpdateChannel(ctx context.Context, in *UpdateChannelRequest, opts ...http.CallOption) (*ChannelInfo, error) {
	var out ChannelInfo
	pattern := "/api/v1/payment/channels/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentUpdateChannel))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}